Add files via upload
# week5_task1: 
Write a program to accept five different numbers by creating a class name that has two friend functions friendfunc1 and friendfunc2 taking two and three numbers in argument respectively and calculate the average of these numbers by passing the object of the class to the friend function

# week5_task2: 
You are working in a construction company and your team is currently tasked with estimating the flooring cost for a rectangular room. The client has provided you with the room's dimensions, and the cost calculator formula based on the area of the floor. Your task is to implement the given solution in C++ to calculate the estimated flooring cost for the given room dimensions.